separated tree view into its own component and added responsive styling based on screen size
