+++ src/components/appbar/AppBar.js
@@ -20,7 +20,7 @@ import HomeIcon from '@mui/icons-material/Home';
 import NavLink from './navLink/navLink';
 import styles from './appbar.module.css';
 import { useCart } from '../../context/CartContext';
-import { SimpleTreeView, TreeItem } from '@mui/x-tree-view';
+import TreeViewDropdown from './components/TreeViewDropdown';
 
 const Search = styled('div')(({ theme }) => ({
   position: 'relative',
@@ -252,31 +252,5 @@ export default function PrimarySearchAppBar() {
   );
 }
 
-function TreeViewDropdown({ onClose }) {
 
-  const categories = {
-    'Gaskets': ['Pre-Cut Gaskets', 'Scanning, Measurement, and Reverse Engineering Services', 'Garlock BLUE_GARD 3000 Pipe', 'Flange Gaskets'],
-    'Tools & Parts': ['Gasket Cutting Machines', 'Replacement Baldes', 'Parts'],
-    // ... more categories
-  };
-  const handleTreeItemClick = (itemId) => {
-}
-
-  return (
-    <SimpleTreeView>
-        {Object.entries(categories).map(([category, subcategories]) => (
-          <TreeItem key={category} itemId={category} label={category}>
-            {subcategories.map(subcategory => (
-              <TreeItem 
-                key={subcategory} 
-                itemId={subcategory} 
-                label={subcategory}
-                onClick={() => handleTreeItemClick(subcategory)}
-              />
-            ))}
-          </TreeItem>
-        ))}
-      </SimpleTreeView>
-  );
-}
