+++ src/components/appbar/appbar.module.css
@@ -1,3 +1,20 @@
 .iconButtonColor {
     color: #fff;
 }
+
+.treeViewDropdown {
+    width: 25%;
+    position: absolute; /* or 'fixed' */
+    z-index: 100; /* Ensure it's above other content */
+    background-color: #fff;
+
+    /* Styles for screens between 600px and 1600px */
+    @media screen and (min-width: 900px) and (max-width: 1600px) {
+        width: 50%;
+    }
+
+    /* Styles for screens smaller than 600px */
+    @media screen and (max-width: 899px) {
+        width: 100%;
+    }
+}
